Add cursor pagination to transactions
